+++ HISTORY.rst
@@ -3,6 +3,12 @@
 History
 -------
 
+0.9.6 (unreleased)
+++++++++++++++++++
+
+* Added more Django 1.10 fixes / cleanups
+* Added python 3.6 compatibility
+
 0.9.5 (2017-02-16)
 ++++++++++++++++++